+++ apps/forge-web/src/layouts/dashboard/sidebar.js
@@ -5,6 +5,7 @@ import { withTheme } from '@material-ui/core/styles';
 
 import { withRouter, Link } from 'react-router-dom';
 import Button from '@material-ui/core/Button';
+import Tooltip from '@material-ui/core/Tooltip';
 
 import Icon8 from '../../components/icon8';
 
@@ -25,7 +26,6 @@ class Sidebar extends React.Component {
 
   // {this.renderMenuItem('/app', 'app', 'Application Management')}
   // {this.renderMenuItem('/tasks', 'tasks', 'Tasks')}
-  // {this.renderMenuItem('/developer', 'developer', 'Developer Tools')}
   // {this.renderMenuItem('/settings', 'settings', 'Settings')}
   render() {
     return (
@@ -33,6 +33,7 @@ class Sidebar extends React.Component {
         <div className="menu-top-items">
           {this.renderMenuItem('/dashboard', 'dashboard', 'Dashboard')}
           {this.renderMenuItem('/node/explorer/txs', 'node', 'Node Management')}
+          {this.renderMenuItem('/developer/query', 'developer', 'Developer Tools')}
         </div>
         <div className="menu-top-items" />
       </MenuItems>
@@ -42,13 +43,15 @@ class Sidebar extends React.Component {
   renderMenuItem(url, name, title) {
     const selected = this.isSelected(url, name);
     return (
-      <MenuItem component={Link} key={url} selected={selected} to={url} title={title}>
-        <Icon8
-          name={images[name]}
-          size={36}
-          color={selected ? '#00c2c4' : this.props.theme.typography.color.main}
-        />
-      </MenuItem>
+      <Tooltip title={title} placement="right">
+        <MenuItem component={Link} key={url} selected={selected} to={url} title={title}>
+          <Icon8
+            name={images[name]}
+            size={36}
+            color={selected ? '#00c2c4' : this.props.theme.typography.color.main}
+          />
+        </MenuItem>
+      </Tooltip>
     );
   }